test(opensearch): fix and refactor InvalidIndexStartupIT
Logic was a little broken since implicit assumption that bpn of the imported business partner is the same as the one in the bogus document in opensearch was wrong. DB wasn't cleared in between tests, so bpn generator counter was not reset. Simplify the test in order to not depend on that assumption at all.